stb ready. some code refactored

+++ build.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
+
 ls cmake &> /dev/null
 RES=$?
 if [[ $RES != 0 ]]; then
 	mkdir cmake
 fi
 cd cmake
-rm -rf *
+if [[ $1 ]]; then 
+	if [[ $1 == "rebuild" ]]; then
+		rm -rf *
+	fi
+fi
 cmake ..
 make
 RES=$?
@@ -14,3 +19,5 @@ if [[ $RES != 0 ]]; then
 else
 	echo -e "\e[1;31mYou can find your binary in $(pwd)\e[0m"
 fi
+
+exit $RES
